Mejoras a los daemon de jabber.
git-svn-id: file:///home/gallir/meneame/svn/src@596 cbe1a1ec-7179-4919-8e32-4551e8279a00

+++ branches/version2/jabber/posts.pl
@@ -226,7 +226,9 @@ sub InMessage
 	}
 	if ( $type ne 'chat' ) {
 		print "Error type '$type' from $from\n";
-		$Users->delete($user);
+		if ( $type eq 'error' ) {
+			$Users->delete($user);
+		}
 		return;
 	}
 	StorePost($user, $body);
